pyproject.toml
[tool.poetry]
name = "python-pdf-parser"
version = "0.1.0"
description = ""
authors = ["nichnikov <a.nichnikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pypdf2 = "^3.0.1"
pdfminer-six = "^20231228"
pdfplumber = "^0.11.1"
pdf2image = "^1.17.0"
pillow = "^10.4.0"
pytesseract = "^0.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
